hashCodeToBytesを移動する
